# docs/site/content/en/openapi/openapi.yaml
...
openapi: 3.0.3
info:
  title: Horreum REST API
  description: "Horreum automated change anomaly detection. For more information,\
    \ please see [https://horreum.hyperfoil.io/](https://horreum.hyperfoil.io/)"
  version: 0.10-SNAPSHOT
...
paths:
...
  /api/experiment/run:
    get:
      tags:
      - Experiment
      operationId: runExperiments
      parameters:
      - name: datasetId
        in: query
        schema:
          format: int32
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExperimentResult'
  /api/experiment/{testId}/profiles:
    get:
      tags:
      - Experiment
      description: Retrieve Experiment Profiles by Test ID
      operationId: profiles
      parameters:
      - name: testId
        in: path
        description: Test ID to retrieve Experiment Profiles for
        required: true
        schema:
          format: int32
          type: integer
        example: 101
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExperimentProfile'
    post:
      tags:
      - Experiment
      description: 'Save new or update existing Experiment Profiles for a Test '
      operationId: addOrUpdateProfile
      parameters:
      - name: testId
        in: path
        description: Test ID to retrieve Experiment Profiles for
        required: true
        schema:
          format: int32
          type: integer
        example: 101
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExperimentProfile'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                format: int32
                type: integer
  /api/experiment/{testId}/profiles/{profileId}:
    delete:
      tags:
      - Experiment
      description: Delete an Experiment Profiles for a Test
      operationId: deleteProfile
      parameters:
      - name: profileId
        in: path
        description: Experiment Profile ID
        required: true
        schema:
          format: int32
          type: integer
        example: 202
      - name: testId
        in: path
        description: Test ID
        required: true
        schema:
          format: int32
          type: integer
        example: 101
      responses:
        "204":
          description: No Content
...
components:
  schemas:
...
    BetterOrWorse:
      enum:
      - BETTER
      - SAME
      - WORSE
      type: string
    ComparisonResult:
      description: Result of performing a Comparison
      type: object
      properties:
        overall:
          allOf:
          - $ref: '#/components/schemas/BetterOrWorse'
          - description: Was the Experiment dataset better or worse than the baseline
              dataset
        experimentValue:
          format: double
          description: Experiment value
          type: number
        baselineValue:
          format: double
          description: Baseline value
          type: number
        result:
          description: The relative difference between the Experiment and Baseline
            Datasets
          type: string
...
    DatasetInfo:
      required:
      - id
      - runId
      - ordinal
      - testId
      type: object
      properties:
        id:
          format: int32
          description: Dataset ID for Dataset
          type: integer
          example: 101
        runId:
          format: int32
          description: Run ID that Dataset relates to
          type: integer
          example: 101
        ordinal:
          format: int32
          description: Ordinal position in ordered list
          type: integer
          example: 2
        testId:
          format: int32
          description: Test ID that Dataset relates to
          type: integer
          example: 103
...
    DatasetLog:
      required:
      - id
      - level
      - timestamp
      - message
      - source
      - testId
      - runId
      - datasetId
      - datasetOrdinal
      type: object
      properties:
        id:
          format: int64
          type: integer
        level:
          format: int32
          type: integer
          nullable: false
        timestamp:
          format: date-time
          type: string
          nullable: false
        message:
          type: string
          nullable: false
        source:
          type: string
          nullable: false
        testId:
          format: int32
          type: integer
        runId:
          format: int32
          type: integer
        datasetId:
          format: int32
          type: integer
        datasetOrdinal:
          format: int32
          type: integer
...
    ExperimentComparison:
      required:
      - model
      - config
      - variableId
      type: object
      properties:
        model:
          description: Name of comparison model
          type: string
          example: relativeDifference
          nullable: false
        config:
          description: Model JSON configuration
          type: string
          nullable: false
        variableId:
          format: int32
          description: Variable ID to run experiment against
          type: integer
          example: 101
          nullable: false
        variableName:
          description: Variable Name to run experiment against
          type: string
          example: Throughput
    ExperimentProfile:
      description: An Experiment Profile defines the labels and filters for the dataset
        and baseline
      required:
      - id
      - name
      - selectorLabels
      - baselineLabels
      - comparisons
      type: object
      properties:
        id:
          format: int32
          description: Experiment Profile unique ID
          type: integer
          example: 101
        name:
          description: Name of Experiment Profile
          type: string
          example: Techempower comparison
          nullable: false
        testId:
          format: int32
          description: Test ID that Experiment Profile relates to
          type: integer
          example: 101
        selectorLabels:
          description: Array of selector labels
          type: array
          items:
            type: string
          example:
          - Framework
        selectorFilter:
          description: Selector filter to apply to Selector label values
          type: string
          example: value => value === 'quarkus-resteasy-reactive-hibernate-reactive'
        baselineLabels:
          description: Array of selector labels for comparison Baseline
          type: array
          items:
            type: string
          example:
          - timestamp
        baselineFilter:
          description: Selector filter to apply to Baseline label values
          type: string
          example: value => value === 1666955225547
        comparisons:
          description: Collection of Experiment Comparisons to run during an Experiment
            evaluation
          type: array
          items:
            $ref: '#/components/schemas/ExperimentComparison'
        extraLabels:
          description: These labels are not used by Horreum but are added to the result
            event and therefore can be used e.g. when firing an Action.
          type: array
          items:
            type: string
    ExperimentResult:
      description: Result of running an Experiment
      type: object
      properties:
        profile:
          allOf:
          - $ref: '#/components/schemas/ExperimentProfile'
          - description: Experiment profile that results relates to
        logs:
          description: A list of log statements recorded while Experiment was evaluated
          type: array
          items:
            $ref: '#/components/schemas/DatasetLog'
        datasetInfo:
          allOf:
          - $ref: '#/components/schemas/DatasetInfo'
          - description: Dataset Info about dataset used for experiment
        baseline:
          description: A list of Dataset Info for experiment baseline(s)
          type: array
          items:
            $ref: '#/components/schemas/DatasetInfo'
        results:
          description: A Map of all comparisons and results evaluated during an Experiment
          type: object
          additionalProperties:
            $ref: '#/components/schemas/ComparisonResult'
        extraLabels:
          type: string
        notify:
          type: boolean
...